feat: Use Zap for logger injection
This changes so that sprout.Logger injects a *zap.Logger. This makes the
library a bit more opinionated but Zap provides a more familiar logging
API. Logr is still supported via sprout.LogrLogger for those cases where
a Logr-bridge is needed.

As part of this change root loggers are no longer injected, so that
fx.Provide can be used instead of fx.Decorate. The new recommended way
to inject a logger in a module is to use fx.Private:

  fx.Provide(sprout.Logger("example"), fx.Private)

A new helper named sprout.SugaredLogger is available if a *zap.SugaredLogger
is wanted instead of a *zap.Logger.