feat: Introduce HealthChecks for adding liveness and readiness checks
This changes from the use of fx.Annotate and fx.Provide to provide
groups to a health check server. The previous approach had a limitation
where injects were not scoped to the module that declared the checks.

This new approach provides sprout.HealthChecks that can be used to add
liveness and readiness checks in a similar way that fx.Lifecycle is used
to add OnStart and OnStop hooks.